requirements.txt
torch>=2.7
transformers>=4.57
peft>=0.11
deepspeed>=0.18
sentence-transformers>=2.7
scikit-learn>=1.5
PyYAML>=6.0
tqdm>=4.66
numpy>=1.26
jsonlines>=4.0
